refactor: rename schemas.ts -> schemaMultipayload.ts
